Fix the prototype of getline on platforms where we need to redefine it (non-GNU)
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@3348 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/xbt/getline.c
@@ -23,7 +23,7 @@
 #include "portable.h"
 
 #ifndef HAVE_GETLINE
-ssize_t getline(char **buf, size_t *n, FILE *stream) {
+long getline(char **buf, size_t *n, FILE *stream) {
    
    int i, ch;
